Remove dead code and move some code around
